work: Local fixes and improvements for case registration
- Fixed Additional Info endpoint table name configuration
- Removed debug code from AdditionalInfoController
- Implemented deficiencies feature with full CRUD support
- Updated classifications and case registration components
- Improved form validation and state management
- Enhanced API endpoints for case registration workflow

+++ src/Backend/BOG.DbModel/Entities/CaseRegistration/RelatedCase.cs
@@ -15,7 +15,7 @@ public class RelatedCase : BaseEntity
     /// <summary>
     /// Related case number (رقم القضية المرتبطة).
     /// </summary>
-    public long CaseNumber { get; set; }
+    public int CaseNumber { get; set; }
 
     /// <summary>
     /// Case year for the related case.
